feat(ci): Add continuous integration workflow
This commit adds a new continuous integration workflow file, ci.yaml, to the project. The workflow is triggered on push and pull requests to the main branch, as well as manually using the workflow_dispatch event. The workflow sets up the environment, installs dependencies using pnpm, and builds the project using Next.js. The pnpm store directory is also configured for caching.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>